reactor: grouping type
