+++ packages/core/src/bindings/index.ts
@@ -1,2 +1,3 @@
 export * from './interfaces';
 export * from './helpers';
+export * from './types';
