+++ packages/core/src/bindings/interfaces/index.ts
@@ -3,5 +3,3 @@ export * from './custom-function-binding';
 export * from './base-function-binding';
 export * from './function-config';
 export * from './triggers';
-export * from './types';
-export * from './bindings-type';
