+++ packages/core/src/bindings/types/index.ts
@@ -1,0 +1,2 @@
+export * from './binding';
+export * from './get-context-bindings';
