+++ packages/core/src/types/union-to-array.ts
@@ -1,0 +1,17 @@
+import type { UnionToIntersection } from 'type-fest';
+
+/**
+ * Ref: https://catchts.com/union-array
+ */
+
+/**
+ * Converts union to overloaded function
+ */
+type UnionToOvlds<U> = UnionToIntersection<U extends any ? (f: U) => void : never>;
+
+type PopUnion<U> = UnionToOvlds<U> extends (a: infer A) => void ? A : never;
+type IsUnion<T> = [T] extends [UnionToIntersection<T>] ? false : true;
+
+export type UnionToArray<T, A extends unknown[] = []> = IsUnion<T> extends true
+  ? UnionToArray<Exclude<T, PopUnion<T>>, [PopUnion<T>, ...A]>
+  : [T, ...A];
